1、数据同步，修改了dnake报错，则不删数据。 2、数据源报错问题。 3、登录优化 4、登录返回，详细地址，业主关系，是否初始化密码

+++ modules/community-synchrodata/community-synchrodata-server/src/main/java/com/mit/community/schedule/BigDataSchedule.java
@@ -56,6 +56,9 @@ public class BigDataSchedule {
     @Transactional(rollbackFor = Exception.class)
     public void countActivePeopleNum (){
         List<String> communityCodeList = clusterCommunityService.listCommunityCodeListByCityName("鹰潭市");
+        if(communityCodeList.isEmpty()){
+            return;
+        }
         activePeopleService.remove();
         communityCodeList.forEach(item -> {
             List<Device> devices = deviceService.listInOrOutByCommunityCode(item, "进");
@@ -79,8 +82,8 @@ public class BigDataSchedule {
     public void  countAgeConstruction (){
         List<String> communityCodeList = clusterCommunityService.listCommunityCodeListByCityName("鹰潭市");
         List<AgeConstruction> ageConstructions = houseHoldService.countAgeConstructionByCommuintyCodeList(communityCodeList);
-        ageConstructionService.remove();
         if(!ageConstructions.isEmpty()){
+            ageConstructionService.remove();
             ageConstructionService.insertBatch(ageConstructions);
         }
     }
